Added pages/routes to pages, simple calendar section
